Add preliminary StubWorkerClient extraction
